fit_model moved to stats
